Abstract out ExecutionsRequestBodyAggregator and fix tests
