+++ metricsaggregator/src/test/java/io/dockstore/metricsaggregator/helper/AggregationHelperTest.java
@@ -1,9 +1,6 @@
 package io.dockstore.metricsaggregator.helper;
 
 import static io.dockstore.metricsaggregator.common.TestUtilities.createValidationExecution;
-import static io.dockstore.openapi.client.model.RunExecution.ExecutionStatusEnum.FAILED_RUNTIME_INVALID;
-import static io.dockstore.openapi.client.model.RunExecution.ExecutionStatusEnum.FAILED_SEMANTIC_INVALID;
-import static io.dockstore.openapi.client.model.RunExecution.ExecutionStatusEnum.SUCCESSFUL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -11,16 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.dockstore.openapi.client.model.AggregatedExecution;
-import io.dockstore.openapi.client.model.Cost;
-import io.dockstore.openapi.client.model.CostMetric;
-import io.dockstore.openapi.client.model.CpuMetric;
-import io.dockstore.openapi.client.model.ExecutionStatusMetric;
-import io.dockstore.openapi.client.model.ExecutionTimeMetric;
 import io.dockstore.openapi.client.model.ExecutionsRequestBody;
-import io.dockstore.openapi.client.model.MemoryMetric;
-import io.dockstore.openapi.client.model.MetricsByStatus;
-import io.dockstore.openapi.client.model.RunExecution;
-import io.dockstore.openapi.client.model.TaskExecutions;
 import io.dockstore.openapi.client.model.ValidationExecution;
 import io.dockstore.openapi.client.model.ValidationStatusMetric;
 import io.dockstore.openapi.client.model.ValidatorInfo;
@@ -33,301 +21,6 @@ import java.util.Optional;
 import org.junit.jupiter.api.Test;
 
 class AggregationHelperTest {
-
-    @Test
-    void testGetAggregatedExecutionStatus() {
-        ExecutionStatusAggregator executionStatusAggregator = new ExecutionStatusAggregator();
-        ExecutionsRequestBody allSubmissions = new ExecutionsRequestBody();
-        Optional<ExecutionStatusMetric> executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isEmpty());
-
-        RunExecution submittedRunExecution = new RunExecution().executionStatus(SUCCESSFUL);
-        allSubmissions = new ExecutionsRequestBody().runExecutions(List.of(submittedRunExecution));
-        executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isPresent());
-        assertEquals(1, executionStatusMetric.get().getCount().get(SUCCESSFUL.toString()).getExecutionStatusCount());
-
-        // Aggregate submissions containing run executions and aggregated metrics
-        AggregatedExecution submittedAggregatedMetrics = new AggregatedExecution();
-        submittedAggregatedMetrics.executionStatusCount(
-                new ExecutionStatusMetric().count(
-                        Map.of(SUCCESSFUL.toString(), new MetricsByStatus().executionStatusCount(10),
-                                FAILED_RUNTIME_INVALID.toString(), new MetricsByStatus().executionStatusCount(1))));
-        allSubmissions = new ExecutionsRequestBody().runExecutions(List.of(submittedRunExecution)).aggregatedExecutions(List.of(submittedAggregatedMetrics));
-        executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isPresent());
-        assertEquals(11, executionStatusMetric.get().getCount().get(SUCCESSFUL.toString()).getExecutionStatusCount());
-        assertEquals(1, executionStatusMetric.get().getCount().get(FAILED_RUNTIME_INVALID.toString()).getExecutionStatusCount());
-        assertNull(executionStatusMetric.get().getCount().get(FAILED_SEMANTIC_INVALID.toString()), "Should be null because the key doesn't exist in the count map");
-
-        // Aggregate submissions containing workflow run executions and task executions
-        submittedRunExecution = new RunExecution().executionStatus(SUCCESSFUL);
-        TaskExecutions taskExecutionsForOneWorkflowRun = new TaskExecutions().taskExecutions(List.of(submittedRunExecution));
-        allSubmissions = new ExecutionsRequestBody().runExecutions(List.of(submittedRunExecution)).taskExecutions(List.of(taskExecutionsForOneWorkflowRun));
-        executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isPresent());
-        assertEquals(2, executionStatusMetric.get().getCount().get(SUCCESSFUL.toString()).getExecutionStatusCount());
-        // Submit one successful workflow execution and a list of task executions where the single task failed
-        taskExecutionsForOneWorkflowRun = new TaskExecutions().taskExecutions(List.of(new RunExecution().executionStatus(FAILED_RUNTIME_INVALID)));
-        allSubmissions = new ExecutionsRequestBody().runExecutions(List.of(submittedRunExecution)).taskExecutions(List.of(taskExecutionsForOneWorkflowRun));
-        executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isPresent());
-        assertEquals(1, executionStatusMetric.get().getCount().get(SUCCESSFUL.toString()).getExecutionStatusCount());
-        assertEquals(1, executionStatusMetric.get().getCount().get(FAILED_RUNTIME_INVALID.toString()).getExecutionStatusCount());
-        // Submit one successful workflow execution and a list of task executions where there are two tasks that failed due to different reasons
-        taskExecutionsForOneWorkflowRun = new TaskExecutions().taskExecutions(
-                List.of(new RunExecution().executionStatus(FAILED_RUNTIME_INVALID),
-                        new RunExecution().executionStatus(FAILED_SEMANTIC_INVALID)));
-        allSubmissions = new ExecutionsRequestBody().runExecutions(List.of(submittedRunExecution)).taskExecutions(List.of(taskExecutionsForOneWorkflowRun));
-        executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isPresent());
-        assertEquals(1, executionStatusMetric.get().getCount().get(SUCCESSFUL.toString()).getExecutionStatusCount());
-        // There should be 1 of either FAILED_RUNTIME_INVALID or FAILED_SEMANTIC_INVALID.
-        assertTrue(executionStatusMetric.get().getCount().containsKey(FAILED_RUNTIME_INVALID.toString()) || executionStatusMetric.get().getCount().containsKey(FAILED_SEMANTIC_INVALID.toString()));
-        // Submit one successful workflow execution and a list of task executions where there are 3 tasks that failed due to different reasons
-        taskExecutionsForOneWorkflowRun = new TaskExecutions().taskExecutions(
-                List.of(new RunExecution().executionStatus(FAILED_RUNTIME_INVALID),
-                        new RunExecution().executionStatus(FAILED_RUNTIME_INVALID),
-                        new RunExecution().executionStatus(FAILED_SEMANTIC_INVALID)));
-        allSubmissions = new ExecutionsRequestBody().runExecutions(List.of(submittedRunExecution)).taskExecutions(List.of(taskExecutionsForOneWorkflowRun));
-        executionStatusMetric = executionStatusAggregator.getAggregatedMetricFromAllSubmissions(allSubmissions);
-        assertTrue(executionStatusMetric.isPresent());
-        assertEquals(1, executionStatusMetric.get().getCount().get(SUCCESSFUL.toString()).getExecutionStatusCount());
-        // There should be one count of FAILED_RUNTIME_INVALID because it's the most frequent failed status in the list of task executions
-        assertEquals(1, executionStatusMetric.get().getCount().get(FAILED_RUNTIME_INVALID.toString()).getExecutionStatusCount());
-    }
-
-    @Test
-    void testGetAggregatedExecutionTime() {
-        ExecutionTimeAggregator executionTimeAggregator = new ExecutionTimeAggregator();
-        List<RunExecution> badExecutions = new ArrayList<>();
-        Optional<ExecutionTimeMetric> executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(badExecutions));
-        assertTrue(executionTimeMetric.isEmpty());
-
-        // Add an execution that doesn't have execution time data
-        badExecutions.add(new RunExecution().executionStatus(SUCCESSFUL));
-        executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(badExecutions));
-        assertTrue(executionTimeMetric.isEmpty());
-
-        // Add an execution with malformed execution time data
-        badExecutions.add(new RunExecution().executionStatus(SUCCESSFUL).executionTime("1 second"));
-        executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(badExecutions));
-        assertTrue(executionTimeMetric.isEmpty());
-
-        // Add an execution with execution time
-        final int timeInSeconds = 10;
-        List<RunExecution> executions = List.of(new RunExecution().executionStatus(SUCCESSFUL).executionTime(String.format("PT%dS", timeInSeconds)));
-        executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(executionTimeMetric.isPresent());
-        assertEquals(timeInSeconds, executionTimeMetric.get().getMinimum());
-        assertEquals(timeInSeconds, executionTimeMetric.get().getMaximum());
-        assertEquals(timeInSeconds, executionTimeMetric.get().getAverage());
-        assertEquals(1, executionTimeMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing run executions and aggregated metrics
-        AggregatedExecution submittedAggregatedMetrics = new AggregatedExecution();
-        ExecutionStatusMetric executionStatusMetric = new ExecutionStatusMetric().putCountItem(SUCCESSFUL.name(),
-                new MetricsByStatus()
-                        .executionStatusCount(1)
-                        .executionTime(new ExecutionTimeMetric()
-                            .minimum(2.0)
-                            .maximum(6.0)
-                            .average(4.0)
-                            .numberOfDataPointsForAverage(2)));
-        submittedAggregatedMetrics.executionStatusCount(executionStatusMetric);
-        executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).aggregatedExecutions(List.of(submittedAggregatedMetrics)));
-        assertTrue(executionTimeMetric.isPresent());
-        assertEquals(2.0, executionTimeMetric.get().getMinimum());
-        assertEquals(10.0, executionTimeMetric.get().getMaximum());
-        assertEquals(6, executionTimeMetric.get().getAverage());
-        assertEquals(3, executionTimeMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing workflow run executions and task executions
-        // Submit a single workflow execution that took 10s and a single task that took 10s
-        executions = List.of(new RunExecution().executionStatus(SUCCESSFUL).executionTime(String.format("PT%dS", timeInSeconds)));
-        executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).taskExecutions(List.of(new TaskExecutions().taskExecutions(executions))));
-        assertTrue(executionTimeMetric.isPresent());
-        assertEquals(timeInSeconds, executionTimeMetric.get().getMinimum());
-        assertEquals(timeInSeconds, executionTimeMetric.get().getMaximum());
-        assertEquals(timeInSeconds, executionTimeMetric.get().getAverage());
-        assertEquals(2, executionTimeMetric.get().getNumberOfDataPointsForAverage()); // There should be 2 data points: 1 for the workflow execution and 1 for the list of tasks
-        // Submit a single workflow execution that took 10s and two tasks that took 10 seconds. This time, dateExecuted is provided
-        RunExecution task1 = new RunExecution().executionStatus(SUCCESSFUL).executionTime(String.format("PT%dS", timeInSeconds));
-        RunExecution task2 = new RunExecution().executionStatus(SUCCESSFUL).executionTime(String.format("PT%dS", timeInSeconds));
-        // The time difference between these two tasks is 10 seconds. When there is more than one task, the duration will be calculated from the dates executed, plus the duration of the last task, which is 10s
-        task1.setDateExecuted("2023-11-09T21:54:10.571285905Z");
-        task2.setDateExecuted("2023-11-09T21:54:20.571285905Z");
-        executionTimeMetric = executionTimeAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).taskExecutions(List.of(new TaskExecutions().taskExecutions(List.of(task1, task2)))));
-        assertTrue(executionTimeMetric.isPresent());
-        assertEquals(10, executionTimeMetric.get().getMinimum(), "The minimum is from the workflow execution");
-        assertEquals(20, executionTimeMetric.get().getMaximum(), "The maximum is from the workflow execution calculated from the two tasks");
-        assertEquals(15, executionTimeMetric.get().getAverage());
-        assertEquals(2, executionTimeMetric.get().getNumberOfDataPointsForAverage()); // There should be 2 data points: 1 for the workflow execution and 1 for the list of tasks
-    }
-
-    @Test
-    void testGetAggregatedCpu() {
-        CpuAggregator cpuAggregator = new CpuAggregator();
-        List<RunExecution> executions = new ArrayList<>();
-        Optional<CpuMetric> cpuMetric = cpuAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(cpuMetric.isEmpty());
-
-        // Add an execution that doesn't have cpu data
-        executions.add(new RunExecution().executionStatus(SUCCESSFUL));
-        cpuMetric = cpuAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(cpuMetric.isEmpty());
-
-        // Add an execution with cpu data
-        final int cpu = 1;
-        executions.add(new RunExecution().executionStatus(SUCCESSFUL).cpuRequirements(cpu));
-        cpuMetric = cpuAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(cpuMetric.isPresent());
-        assertEquals(cpu, cpuMetric.get().getMinimum());
-        assertEquals(cpu, cpuMetric.get().getMaximum());
-        assertEquals(cpu, cpuMetric.get().getAverage());
-        assertEquals(1, cpuMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing run executions and aggregated metrics
-        AggregatedExecution submittedAggregatedMetrics = new AggregatedExecution();
-        ExecutionStatusMetric executionStatusMetric = new ExecutionStatusMetric().putCountItem(SUCCESSFUL.name(),
-                new MetricsByStatus()
-                        .executionStatusCount(1)
-                        .cpu(new CpuMetric()
-                                .minimum(2.0)
-                                .maximum(6.0)
-                                .average(4.0)
-                                .numberOfDataPointsForAverage(2)));
-        submittedAggregatedMetrics.executionStatusCount(executionStatusMetric);
-        cpuMetric = cpuAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).aggregatedExecutions(List.of(submittedAggregatedMetrics)));
-        assertTrue(cpuMetric.isPresent());
-        assertEquals(1.0, cpuMetric.get().getMinimum());
-        assertEquals(6.0, cpuMetric.get().getMaximum());
-        assertEquals(3, cpuMetric.get().getAverage());
-        assertEquals(3, cpuMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing workflow run executions and task executions
-        executions = List.of(new RunExecution().executionStatus(SUCCESSFUL).cpuRequirements(cpu));
-        // Two task executions with different CPU requirements. The workflow execution calculated from these tasks should take the highest cpuRequirement from the tasks
-        TaskExecutions taskExecutions = new TaskExecutions().taskExecutions(List.of(
-                new RunExecution().executionStatus(SUCCESSFUL).cpuRequirements(1),
-                new RunExecution().executionStatus(SUCCESSFUL).cpuRequirements(4)));
-        cpuMetric = cpuAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).taskExecutions(List.of(taskExecutions)));
-        assertTrue(cpuMetric.isPresent());
-        assertEquals(1.0, cpuMetric.get().getMinimum());
-        assertEquals(4.0, cpuMetric.get().getMaximum());
-        assertEquals(2.5, cpuMetric.get().getAverage());
-        assertEquals(2, cpuMetric.get().getNumberOfDataPointsForAverage()); // Two data points: 1 from the workflow execution and 1 for the list of tasks
-    }
-
-    @Test
-    void testGetAggregatedMemory() {
-        MemoryAggregator memoryAggregator = new MemoryAggregator();
-        List<RunExecution> executions = new ArrayList<>();
-        Optional<MemoryMetric> memoryMetric = memoryAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(memoryMetric.isEmpty());
-
-        // Add an execution that doesn't have memory data
-        executions.add(new RunExecution().executionStatus(SUCCESSFUL));
-        memoryMetric = memoryAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(memoryMetric.isEmpty());
-
-        // Add an execution with memory data
-        Double memoryInGB = 2.0;
-        executions.add(new RunExecution().executionStatus(SUCCESSFUL).memoryRequirementsGB(memoryInGB));
-        memoryMetric = memoryAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(memoryMetric.isPresent());
-        assertEquals(memoryInGB, memoryMetric.get().getMinimum());
-        assertEquals(memoryInGB, memoryMetric.get().getMaximum());
-        assertEquals(memoryInGB, memoryMetric.get().getAverage());
-        assertEquals(1, memoryMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing run executions and aggregated metrics
-        AggregatedExecution submittedAggregatedMetrics = new AggregatedExecution();
-        ExecutionStatusMetric executionStatusMetric = new ExecutionStatusMetric().putCountItem(SUCCESSFUL.name(),
-                new MetricsByStatus()
-                        .executionStatusCount(1)
-                        .memory(new MemoryMetric()
-                                .minimum(2.0)
-                                .maximum(6.0)
-                                .average(4.0)
-                                .numberOfDataPointsForAverage(2)));
-        submittedAggregatedMetrics.executionStatusCount(executionStatusMetric);
-        memoryMetric = memoryAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).aggregatedExecutions(List.of(submittedAggregatedMetrics)));
-        assertTrue(memoryMetric.isPresent());
-        assertEquals(2.0, memoryMetric.get().getMinimum());
-        assertEquals(6.0, memoryMetric.get().getMaximum());
-        assertEquals(3.333333333333333, memoryMetric.get().getAverage());
-        assertEquals(3, memoryMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing workflow run executions and task executions
-        executions = List.of(new RunExecution().executionStatus(SUCCESSFUL).memoryRequirementsGB(2.0));
-        // Two task executions with different memory requirements. The workflow execution calculated from these tasks should take the highest memoryRequirementsGB from the tasks
-        TaskExecutions taskExecutions = new TaskExecutions().taskExecutions(List.of(
-                new RunExecution().executionStatus(SUCCESSFUL).memoryRequirementsGB(2.0),
-                new RunExecution().executionStatus(SUCCESSFUL).memoryRequirementsGB(4.0)));
-        memoryMetric = memoryAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).taskExecutions(List.of(taskExecutions)));
-        assertTrue(memoryMetric.isPresent());
-        assertEquals(2.0, memoryMetric.get().getMinimum());
-        assertEquals(4.0, memoryMetric.get().getMaximum());
-        assertEquals(3.0, memoryMetric.get().getAverage());
-        assertEquals(2, memoryMetric.get().getNumberOfDataPointsForAverage()); // Two data points: 1 from the workflow execution and 1 for the list of tasks
-
-    }
-
-    @Test
-    void testGetAggregatedCost() {
-        CostAggregator costAggregator = new CostAggregator();
-        List<RunExecution> executions = new ArrayList<>();
-        Optional<CostMetric> costMetric = costAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(costMetric.isEmpty());
-
-        // Add an execution that doesn't have cost data
-        executions.add(new RunExecution().executionStatus(SUCCESSFUL));
-        costMetric = costAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(costMetric.isEmpty());
-
-        // Add an execution with cost data
-        Double costInUSD = 2.00;
-        executions.add(new RunExecution().executionStatus(SUCCESSFUL).cost(new Cost().value(costInUSD)));
-        costMetric = costAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions));
-        assertTrue(costMetric.isPresent());
-        assertEquals(costInUSD, costMetric.get().getMinimum());
-        assertEquals(costInUSD, costMetric.get().getMaximum());
-        assertEquals(costInUSD, costMetric.get().getAverage());
-        assertEquals(1, costMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing run executions and aggregated metrics
-        AggregatedExecution submittedAggregatedMetrics = new AggregatedExecution();
-        ExecutionStatusMetric executionStatusMetric = new ExecutionStatusMetric().putCountItem(SUCCESSFUL.name(),
-                new MetricsByStatus()
-                        .executionStatusCount(1)
-                        .cost(new CostMetric()
-                                .minimum(2.0)
-                                .maximum(6.0)
-                                .average(4.0)
-                                .numberOfDataPointsForAverage(2)));
-        submittedAggregatedMetrics.executionStatusCount(executionStatusMetric);
-        costMetric = costAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).aggregatedExecutions(List.of(submittedAggregatedMetrics)));
-        assertTrue(costMetric.isPresent());
-        assertEquals(2.0, costMetric.get().getMinimum());
-        assertEquals(6.0, costMetric.get().getMaximum());
-        assertEquals(3.333333333333333, costMetric.get().getAverage());
-        assertEquals(3, costMetric.get().getNumberOfDataPointsForAverage());
-
-        // Aggregate submissions containing workflow run executions and task executions
-        executions = List.of(new RunExecution().executionStatus(SUCCESSFUL).cost(new Cost().value(2.00)));
-        // Two task executions with different memory requirements. The workflow execution calculated from these tasks should have the sum of the cost of all tasks
-        TaskExecutions taskExecutions = new TaskExecutions().taskExecutions(List.of(
-                new RunExecution().executionStatus(SUCCESSFUL).cost(new Cost().value(2.00)),
-                new RunExecution().executionStatus(SUCCESSFUL).cost(new Cost().value(4.00))
-        ));
-        costMetric = costAggregator.getAggregatedMetricFromAllSubmissions(new ExecutionsRequestBody().runExecutions(executions).taskExecutions(List.of(taskExecutions)));
-        assertTrue(costMetric.isPresent());
-        assertEquals(2.0, costMetric.get().getMinimum());
-        assertEquals(6.0, costMetric.get().getMaximum()); // The max is the cost of the two tasks summed together
-        assertEquals(4.0, costMetric.get().getAverage());
-        assertEquals(2, costMetric.get().getNumberOfDataPointsForAverage()); // Two data points: 1 from the workflow execution and 1 for the list of tasks
-    }
 
     @Test
     void testGetAggregatedValidationStatus() {
